Graphql (#41)
* Prototype for GraphQL API

GraphQL is used to receive number of watchers, total number of commits,
and number of people who star the repository in order to build the
correctness metric. Data is converted from string to json as the data
type specified as CorrectnessFactors which resembles the GraphQL query.

Should throw better errors.

* linted graphql.go

* Conversion of NPM url to Github url done (not well tested)

NPMtoGithubUrl takes in a npm url and converts it to a github url
through extraction of json data from npm registry and parsing of
strings until github url is found. Pretty inneficient and may be a
cause of lag because of huge json to string conversions and parsing.

Small fixes added to graphql.go to make it run.

* rework of graphql api. token removed. use of npm command to convert npm to github url

* first revision

* test

* remove ls

* cleanup

* Conversion of NPM url to Github url done (not well tested)

NPMtoGithubUrl takes in a npm url and converts it to a github url
through extraction of json data from npm registry and parsing of
strings until github url is found. Pretty inneficient and may be a
cause of lag because of huge json to string conversions and parsing.

Small fixes added to graphql.go to make it run.


* Conversion of NPM url to Github url done (not well tested)

NPMtoGithubUrl takes in a npm url and converts it to a github url
through extraction of json data from npm registry and parsing of
strings until github url is found. Pretty inneficient and may be a
cause of lag because of huge json to string conversions and parsing.

Small fixes added to graphql.go to make it run.

* rework of graphql api. token removed. use of npm command to convert npm to github url

---------

Co-authored-by: Mimi Chon <chonm@purdue.edu>
Co-authored-by: Ben Schwartz <schwartzb.2019@gmail.com>